// src/app/shared/@ui-components/admin-filter-modal/models/filter.ts
export interface Filter {
    startDate?: string;
    endDate?: string;
    dayDate?: string;
    monthYear?: string;
    statusIdList?: string[];
    administrativeLevels?: string[];
    employeeId?: string;
    minimum?: string;
    leaveRegulationId?: string[];
    allowanceId?: string[];
    fullDayId?: string[];
    partialPermissionTypeId?: string[];

}